refactor: move models into their folder
