+++ source/javascripts/core/models/BitriseYml.schema.ts
