+++ source/javascripts/core/models/MachineType.ts
