+++ source/javascripts/core/models/Pipeline.ts
