+++ source/javascripts/core/models/Secret.ts
