+++ source/javascripts/core/models/Stack.ts
