+++ source/javascripts/core/models/Stage.ts
