+++ source/javascripts/core/models/TriggerMap.ts
